refactor: started on creating transformatoins
