+++ cognite/neat/rules/transformers/_base.py
@@ -1,4 +1,5 @@
 from abc import ABC, abstractmethod
+from collections.abc import MutableSequence
 from typing import Generic, TypeVar
 
 from cognite.neat.rules._shared import Rules
@@ -12,4 +13,12 @@ class RulesTransformer(ABC, Generic[T_RulesIn, T_RulesOut]):
 
     @abstractmethod
     def transform(self, rules: T_RulesIn) -> T_RulesOut:
+        """Transform the input rules into the output rules."""
         raise NotImplementedError()
+
+
+class RulesPipeline(list, MutableSequence[RulesTransformer], Generic[T_RulesIn, T_RulesOut]):
+    def run(self, rules: T_RulesIn) -> T_RulesOut:
+        for transformer in self:
+            rules = transformer.transform(rules)
+        return rules  # type: ignore[return-value]
