+++ tests/tests_unit/rules/test_models/test_asset_rules.py
@@ -8,6 +8,7 @@ from cognite.neat.issues.errors import NeatValueError, PropertyDefinitionError
 from cognite.neat.rules.models import AssetRules, InformationRules
 from cognite.neat.rules.models.data_types import DataType
 from cognite.neat.rules.models.entities import AssetEntity, ClassEntity, RelationshipEntity
+from cognite.neat.rules.transformers import AssetToInformation, InformationToAsset, RulesPipeline
 
 
 def case_asset_relationship():
@@ -213,10 +214,10 @@ class TestAssetRules:
         assert AssetRules.model_validate(rules).properties.data[0].implementation == expected_exception
 
     def test_conversion_between_roles(self, david_rules: InformationRules) -> None:
-        asset_rules = david_rules.as_asset_architect_rules()
-        information_rules = asset_rules.as_information_rules()
+        pipeline = RulesPipeline[InformationRules, InformationRules]([InformationToAsset(), AssetToInformation()])
+        recreated = pipeline.run(david_rules)
 
-        assert asset_rules.model_dump() == information_rules.as_asset_architect_rules().model_dump()
+        assert recreated.model_dump() == david_rules.model_dump()
 
     @pytest.mark.parametrize("invalid_rules, expected_exception", list(case_circular_dependency()))
     def test_circular_dependency(self, invalid_rules: dict[str, dict[str, Any]], expected_exception: NeatError) -> None:
